Prevent test imports during production builds of images
